python/requirements.txt
# Web框架
Flask==3.0.2
Flask-Session==0.7.0

# 微信支付相关
pycryptodome==3.20.0
requests==2.31.0

# 二维码生成
qrcode==7.4.2
Pillow==10.2.0

# 环境变量
python-dotenv==1.0.1

# 开发工具
black==24.2.0
flake8==7.0.0
mypy==1.8.0 
cachelib
loguru
cryptography
pyOpenSSL
Pillow